re: move validation methods to struct file
